testing de funciones con cupos

+++ src/main/java/ar/edu/unlam/tallerweb1/domain/funcion/RepositorioFuncion.java
@@ -13,6 +13,6 @@ public interface RepositorioFuncion {
 
 	List<Funcion> getCinesDeUnaPelicula(Long pelicula);
 	
-	Long getCantidadAsientosOcupados(Long funcion);
+	Integer getCantidadAsientosOcupados(Long funcion);
 	
 }
